requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
diff-match-patch==20230430
Django==5.1.1
django-cotton==1.0.8
django-import-export==4.1.1
django-unfold==0.39.0
djangorestframework==3.15.2
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
pathspec==0.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
regex==2024.9.11
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
tqdm==4.66.5
